Modified the kernel algorithm and added a truncation variable
Changed the kernel algorithm to be more compatible with Peter Schmid's
DMD algorithm.  It also allows the possiblility of more heavily
truncated approximations.  It is unclear whether or not this is
a good thing conceptually, but it is useful in practice.

Also added an optional rcond argument to DMD and KDMD, which enable
more heavily truncated approximations to be computed.  The current
code still appends many of the 0 eigenvalues...these should likely
simply be neglected.